[pages][xs]: mv blog.tsx to blog/index.tsx
